Clean up test user fixtures and fix tests

+++ datajunction-server/tests/api/tags_test.py
@@ -26,7 +26,8 @@ class TestTags:
         yield  # Testing happens
         # Teardown: remove any data from database (even data not created by this session)
         for table in Base.metadata.tables.keys():
-            await module__session.execute(text(f'TRUNCATE TABLE "{table}" CASCADE'))
+            if table != "users":
+                await module__session.execute(text(f'TRUNCATE TABLE "{table}" CASCADE'))
         await module__session.commit()
 
     async def create_tag(self, module__client: AsyncClient):
@@ -134,6 +135,16 @@ class TestTags:
         """
         response = await self.create_tag(module__client)
         assert response.status_code == 201
+        response = await module__client.post(
+            "/tags/",
+            json={
+                "name": "sales_report",
+                "display_name": "Sales Report",
+                "description": "All metrics for sales",
+                "tag_type": "group",
+                "tag_metadata": {},
+            },
+        )
 
         # Trying updating the tag
         response = await module__client.patch(
